refactor(modal): move modal file to features and modify its related import
